deband.glsl: completely rework/revamp
This expands the algorithm with the capability to perform multiple
iterations, thus (hopefully) improving the debanding quality at both
large and small scales.

This deprecates deband8 (which was basically a worse version of
ITERATIONS=2), and debandpre was deleted since it just wasn't very
useful in general.

This also adds more documentation and internal code separation.